Fix async dependency tests
