+++ agent/build.gradle
@@ -93,6 +93,7 @@ dependencies {
 
     compile "org.glowroot.xyzzy:xyzzy-engine:$xyzzyVersion"
     compile "org.glowroot.xyzzy:xyzzy-instrumentation-api:$xyzzyVersion"
+    runtime "org.glowroot.xyzzy.instrumentation:instrumentation-executor:$xyzzyVersion"
     runtime "org.glowroot.xyzzy.instrumentation:instrumentation-servlet:$xyzzyVersion"
     runtime "org.glowroot.xyzzy.instrumentation:instrumentation-spring:$xyzzyVersion" // need for operation name
     runtime "org.glowroot.xyzzy.instrumentation:instrumentation-redis:$xyzzyVersion"
