+++ test/smoke/testApps/SpringBootTest/src/main/java/com/SpringBootApp.java
@@ -4,25 +4,40 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.support.SpringBootServletInitializer;
+import org.springframework.context.annotation.Bean;
 import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 
 @SpringBootApplication
 @EnableAsync
 public class SpringBootApp extends SpringBootServletInitializer {
 
-	public SpringBootApp() {
-		super();
+    public SpringBootApp() {
+        super();
 
-		// This lets tomcat handle error and hence filter catches exception.
-		// Disables Springboot error handling which prevents response from propagating up.
-		// See: https://github.com/spring-projects/spring-boot/commit/6381a07c71310c56dc29cf99709adf5fe6e6406a
-		setRegisterErrorPageFilter(false);
-	}
-	@Override
-	protected  SpringApplicationBuilder configure(SpringApplicationBuilder applicationBuilder) {
-		return applicationBuilder.sources(SpringBootApp.class);
-	}
-  public static void main(String[] args) {
-	  SpringApplication.run(SpringBootApp.class, args);
-  }
+        // This lets tomcat handle error and hence filter catches exception.
+        // Disables Springboot error handling which prevents response from propagating up.
+        // See: https://github.com/spring-projects/spring-boot/commit/6381a07c71310c56dc29cf99709adf5fe6e6406a
+        setRegisterErrorPageFilter(false);
+    }
+
+    @Override
+    protected SpringApplicationBuilder configure(SpringApplicationBuilder applicationBuilder) {
+        return applicationBuilder.sources(SpringBootApp.class);
+    }
+
+    @Bean
+    public ThreadPoolTaskExecutor taskExecutor() {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+        executor.setCorePoolSize(2);
+        executor.setMaxPoolSize(2);
+        executor.setQueueCapacity(500);
+        executor.setThreadNamePrefix("Async-");
+        executor.initialize();
+        return executor;
+    }
+
+    public static void main(String[] args) {
+        SpringApplication.run(SpringBootApp.class, args);
+    }
 }
